refactor: Refactor components by handling styled components #31
