+++ client/package.json
@@ -21,7 +21,6 @@
     "react-router-dom": "^6.3.0",
     "redux": "^4.2.0",
     "styled-components": "^5.3.5",
-    "styled-jsx": "^5.0.2",
     "use-interval": "^1.4.0"
   },
   "devDependencies": {
