+++ client/src/main.tsx
@@ -1,34 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
+import { GlobalStyle } from './others/cssLibrary';
 import { store } from './others/store';
 import Router from './Router';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <>
-    <Provider store={store}>
-      <Router />
-    </Provider>
-    <style jsx global='true'>{`
-      * {
-        margin: 0;
-        padding: 0;
-        box-sizing: border-box;
-        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
-          'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-      }
-      html,
-      body,
-      #root,
-      #root > div {
-        width: 100%;
-      }
-      #root {
-        display: flex;
-        padding-top: 70px;
-      }
-    `}</style>
-  </>,
+  <Provider store={store}>
+    <GlobalStyle />
+    <Router />
+  </Provider>,
 );
